extract raylet registration
