+++ src/ray/core_worker/core_worker.cc
@@ -40,6 +40,7 @@
 #include "ray/util/util.h"
 
 using json = nlohmann::json;
+using MessageType = ray::protocol::MessageType;
 
 namespace ray::core {
 
@@ -251,6 +252,53 @@ void TaskCounter::UnsetMetricStatus(const std::string &func_name,
   }
 }
 
+Status CoreWorker::RegisterWorkerToWorkerPool(raylet::RayletConnection &conn,
+                                              const WorkerID &worker_id,
+                                              rpc::WorkerType worker_type,
+                                              const JobID &job_id,
+                                              int runtime_env_hash,
+                                              const Language &language,
+                                              const std::string &ip_address,
+                                              const std::string &serialized_job_config,
+                                              const StartupToken &startup_token,
+                                              NodeID *raylet_id,
+                                              int *port) {
+  flatbuffers::FlatBufferBuilder fbb;
+  // TODO(suquark): Use `WorkerType` in `common.proto` without converting to int.
+  auto message =
+      protocol::CreateRegisterClientRequest(fbb,
+                                            static_cast<int>(worker_type),
+                                            to_flatbuf(fbb, worker_id),
+                                            getpid(),
+                                            startup_token,
+                                            to_flatbuf(fbb, job_id),
+                                            runtime_env_hash,
+                                            language,
+                                            fbb.CreateString(ip_address),
+                                            /*port=*/0,
+                                            fbb.CreateString(serialized_job_config));
+  fbb.Finish(message);
+  // Register the process ID with the raylet.
+  // NOTE(swang): If raylet exits and we are registered as a worker, we will get killed.
+  std::vector<uint8_t> reply;
+  auto request_status = conn.AtomicRequestReply(
+      MessageType::RegisterClientRequest, MessageType::RegisterClientReply, &reply, &fbb);
+  if (!request_status.ok()) {
+    return Status(request_status.code(),
+                  std::string("[RayletClient] Unable to register worker with raylet. ") +
+                      request_status.message());
+  }
+  auto reply_message = flatbuffers::GetRoot<protocol::RegisterClientReply>(reply.data());
+  bool success = reply_message->success();
+  if (!success) {
+    return Status::Invalid(string_from_flatbuf(*reply_message->failure_reason()));
+  }
+
+  *raylet_id = NodeID::FromBinary(reply_message->raylet_id()->str());
+  *port = reply_message->port();
+  return Status::OK();
+}
+
 CoreWorker::CoreWorker(CoreWorkerOptions options, const WorkerID &worker_id)
     : options_(std::move(options)),
       get_call_site_(RayConfig::instance().record_ref_creation_sites()
@@ -349,10 +397,6 @@ CoreWorker::CoreWorker(CoreWorkerOptions options, const WorkerID &worker_id)
 #endif
   io_thread_ = boost::thread(io_thread_attrs, [this]() { RunIOService(); });
 
-  Status raylet_client_status;
-  NodeID local_raylet_id;
-  int assigned_port;
-
   if (options_.worker_type == WorkerType::DRIVER &&
       !options_.serialized_job_config.empty()) {
     // Driver populates the job config via initialization.
@@ -362,32 +406,33 @@ CoreWorker::CoreWorker(CoreWorkerOptions options, const WorkerID &worker_id)
     worker_context_.MaybeInitializeJobInfo(worker_context_.GetCurrentJobID(), job_config);
   }
 
-  local_raylet_client_ =
-      std::make_shared<raylet::RayletClient>(io_service_,
-                                             std::move(grpc_client),
-                                             options_.raylet_socket,
-                                             GetWorkerID(),
-                                             options_.worker_type,
-                                             worker_context_.GetCurrentJobID(),
-                                             options_.runtime_env_hash,
-                                             options_.language,
-                                             options_.node_ip_address,
-                                             &raylet_client_status,
-                                             &local_raylet_id,
-                                             &assigned_port,
-                                             options_.serialized_job_config,
-                                             options_.startup_token);
-
+  auto raylet_conn = std::make_unique<raylet::RayletConnection>(
+      io_service_, options_.raylet_socket, /*num_retries=*/-1, /*timeout=*/-1);
+  NodeID local_raylet_id;
+  int assigned_port = 0;
+  Status raylet_client_status =
+      RegisterWorkerToWorkerPool(*raylet_conn,
+                                 GetWorkerID(),
+                                 options_.worker_type,
+                                 worker_context_.GetCurrentJobID(),
+                                 options_.runtime_env_hash,
+                                 options_.language,
+                                 options_.node_ip_address,
+                                 options_.serialized_job_config,
+                                 options_.startup_token,
+                                 &local_raylet_id,
+                                 &assigned_port);
   if (!raylet_client_status.ok()) {
     // Avoid using FATAL log or RAY_CHECK here because they may create a core dump file.
     RAY_LOG(ERROR).WithField(worker_id)
         << "Failed to register worker to Raylet: " << raylet_client_status;
     QuickExit();
   }
+  RAY_CHECK_GE(assigned_port, 0);
 
+  local_raylet_client_ = std::make_shared<raylet::RayletClient>(
+      std::move(raylet_conn), std::move(grpc_client), GetWorkerID());
   connected_ = true;
-
-  RAY_CHECK(assigned_port >= 0);
 
   // Start RPC server after all the task receivers are properly initialized and we have
   // our assigned port from the raylet.
