+++ src/ray/raylet_client/raylet_client.h
@@ -24,6 +24,7 @@
 #include "ray/common/client_connection.h"
 #include "ray/common/status.h"
 #include "ray/common/task/task_spec.h"
+#include "ray/raylet_client/raylet_connection.h"
 #include "ray/rpc/node_manager/node_manager_client.h"
 #include "ray/util/process.h"
 #include "src/ray/protobuf/common.pb.h"
@@ -38,7 +39,6 @@ using ray::WorkerID;
 
 using ray::Language;
 
-using MessageType = ray::protocol::MessageType;
 // Maps from resource name to its allocation.
 using ResourceMappingType =
     std::unordered_map<std::string, std::vector<std::pair<int64_t, double>>>;
@@ -56,7 +56,7 @@ class PinObjectsInterface {
       const ObjectID &generator_id,
       const ray::rpc::ClientCallback<ray::rpc::PinObjectIDsReply> &callback) = 0;
 
-  virtual ~PinObjectsInterface(){};
+  virtual ~PinObjectsInterface() = default;
 };
 
 /// Interface for leasing workers. Abstract for testing.
@@ -256,45 +256,11 @@ class RayletClientInterface : public PinObjectsInterface,
 
 namespace raylet {
 
-class RayletConnection {
- public:
-  /// Connect to the raylet.
-  ///
-  /// \param raylet_socket The name of the socket to use to connect to the raylet.
-  /// \param worker_id A unique ID to represent the worker.
-  /// \param is_worker Whether this client is a worker. If it is a worker, an
-  ///        additional message will be sent to register as one.
-  /// \param job_id The ID of the driver. This is non-nil if the client is a
-  ///        driver.
-  /// \return The connection information.
-  RayletConnection(instrumented_io_context &io_service,
-                   const std::string &raylet_socket,
-                   int num_retries,
-                   int64_t timeout);
-
-  ray::Status WriteMessage(MessageType type,
-                           flatbuffers::FlatBufferBuilder *fbb = nullptr);
-
-  ray::Status AtomicRequestReply(MessageType request_type,
-                                 MessageType reply_type,
-                                 std::vector<uint8_t> *reply_message,
-                                 flatbuffers::FlatBufferBuilder *fbb = nullptr);
-
- private:
-  /// Shutdown the raylet if the local connection is disconnected.
-  void ShutdownIfLocalRayletDisconnected(const Status &status);
-  /// The connection to raylet.
-  std::shared_ptr<ServerConnection> conn_;
-  /// A mutex to protect stateful operations of the raylet client.
-  std::mutex mutex_;
-  /// A mutex to protect write operations of the raylet client.
-  std::mutex write_mutex_;
-};
-
 class RayletClient : public RayletClientInterface {
  public:
   /// Connect to the raylet.
   ///
+  /// \param raylet_conn connection to raylet.
   /// \param grpc_client gRPC client to the raylet.
   /// \param raylet_socket The name of the socket to use to connect to the raylet.
   /// \param worker_id A unique ID to represent the worker.
@@ -314,20 +280,9 @@ class RayletClient : public RayletClientInterface {
   /// provided by driver will be passed to Raylet.
   /// \param startup_token The startup token of the process assigned to
   /// it during startup as a command line argument.
-  RayletClient(instrumented_io_context &io_service,
+  RayletClient(std::unique_ptr<RayletConnection> raylet_conn,
                std::shared_ptr<ray::rpc::NodeManagerWorkerClient> grpc_client,
-               const std::string &raylet_socket,
-               const WorkerID &worker_id,
-               rpc::WorkerType worker_type,
-               const JobID &job_id,
-               const int &runtime_env_hash,
-               const Language &language,
-               const std::string &ip_address,
-               Status *status,
-               NodeID *raylet_id,
-               int *port,
-               const std::string &serialized_job_config,
-               StartupToken startup_token);
+               const WorkerID &worker_id);
 
   /// Connect to the raylet via grpc only.
   ///
